Revert sample game changes and include colour blending samples
